显示时间,动态加载导出日期,获取k线数据添加失败重试,修复日志打印中文乱码,延长年化波动率保存时间为90天,修复除数为0问题,保留小数统一改为8位,导出年化波动率日期改为手动选择,添加每天12点定时刷新波动率数据

+++ storage.py
@@ -1,7 +1,9 @@
 import os
 import sqlite3
 import pprint
+import time
 import datetime as dt
+from collections import deque
 
 import pandas as pd
 from tkinter import messagebox
@@ -19,7 +21,8 @@ wave_coef_year = 19.1
 
 class DB:
     def __init__(self):
-        self.data_clean_timespan = 30
+        self.data_clean_timespan = 90
+        self.state = 'ok'
         self.wave_rate_col_name_map = {
             'timestamp': '日期',
             'coin_type': '币种名称',
@@ -105,20 +108,28 @@ class DB:
         '''
         self.execute(cmd)
 
-    def init_wave_data(self):
-        row_num = self.execute('select count(*) from wave_rate;')[0][0]
-        if row_num == 0:
-            logger.info('table wave_rate is empty, try to initialize')
-            kline_data = data.get_kline_data()
-            self.handle_kline_data(kline_data)
-            cmd = ('insert into wave_rate (timestamp,begin_price,max_price,min_price,last_price,coin_type,daily_wave,avg_wave,wave_sum,daily_wave_rate,wave_rate_year) values ' +
-                   ','.join([('(%s,%s,%s,%s,%s,"%s",%s,%s,%s,%s,%s)' % tuple(v)) for _, values in kline_data.items() for v in values]) +
+    def insert_wave_rate_batch(self, data):
+        cmd = ('insert into wave_rate (timestamp,begin_price,max_price,min_price,last_price,coin_type,daily_wave,avg_wave,wave_sum,daily_wave_rate,wave_rate_year) values ' +
+                   ','.join([('(%s,%s,%s,%s,%s,"%s",%s,%s,%s,%s,%s)' % tuple(v)) for _, values in data.items() for v in values]) +
                    ';')
-            self.execute(cmd)
-            messagebox.showinfo('提示', '初始化数据完成')
-        else:
-            logger.info('table wave_rate not empty. try to update new data')
-            pass
+        self.execute(cmd)
+
+    def init_wave_data(self):
+        try:
+            self.state = 'initializing'
+            row_num = self.execute('select count(*) from wave_rate;')[0][0]
+            if row_num == 0:
+                logger.info('table wave_rate is empty, try to initialize')
+                kline_data = data.get_kline_data()
+                self.handle_kline_data(kline_data)
+                self.insert_wave_rate_batch(kline_data)
+            else:
+                logger.info('table wave_rate not empty. try to update new data')
+                pass
+        except Exception as e:
+            logger.exception('init wave_data error, reset state to ok: %s', e)
+        finally:
+            self.state = 'ok'
 
     @staticmethod
     def handle_kline_data(kline_data):
@@ -137,9 +148,9 @@ class DB:
                     v.extend(['null'] * 4)
                 else:
                     wave_sum = sum([v_tmp[6] for v_tmp in values[index - 14: index]])
-                    avg_wave = round(wave_sum / wave_coef, ndigits=6)
-                    daily_wave_rate = round(avg_wave / float(v[4]) * 100, ndigits=6)
-                    wave_rate_year = round(daily_wave_rate * wave_coef_year, ndigits=6)
+                    avg_wave = round(wave_sum / wave_coef, ndigits=8)
+                    daily_wave_rate = round(avg_wave / max(float(v[4]), 1e-8) * 100, ndigits=8)
+                    wave_rate_year = round(daily_wave_rate * wave_coef_year, ndigits=8)
                     v.extend([avg_wave, wave_sum, daily_wave_rate, wave_rate_year])
 
     def get_newest_date(self):
@@ -159,20 +170,38 @@ class DB:
         return rst
 
     def clean_wave_rate_old_data(self):
-        newest_date = self.execute('select max(timestamp) from wave_rate;')[0][0]
+        resp = self.execute('select max(timestamp) from wave_rate;')
+        if len(resp) == 0 or len(resp[0]) == 0:
+            return
+        newest_date = resp[0][0]
         date_clean = int(newest_date) - 24 * 3600 * self.data_clean_timespan
         self.execute('delete from wave_rate where timestamp < %s' % date_clean)
         logger.info('success clean wave rate data before date %s', dt.datetime.fromtimestamp(date_clean).strftime('%Y-%m-%d'))
 
-    def export_data(self, export_data_vars):
-        self.clean_wave_rate_old_data()
+    def export_data(self, export_data_vars, wave_date_combobox):
+        if self.state == 'initializing':
+            logger.error('data is initializing, cannot export data')
+            messagebox.showinfo('提示', '正在初始化数据, 无法导出数据')
+            return
         if export_data_vars['年化波动率']:
+            wave_rate_date = wave_date_combobox.get()
+            if not wave_rate_date:
+                logger.error('export wave_rate data need fill wave rate date.')
+                messagebox.showinfo('提示', '导出数据需要填写年化波动率日期')
+                return
+            self.clean_wave_rate_old_data()
             df = self.execute_df('select * from wave_rate;')
-            max_date = df['timestamp'].max()
-            table_name = '年华波动率数据%s.xlsx' % dt.datetime.now().strftime('%Y-%m-%d')
+            target_date = dt.datetime.strptime(wave_rate_date, '%Y-%m-%d')
+            target_timestamp = target_date.timestamp()
+            # max_date = df['timestamp'].max()
+            table_name = '年华波动率数据%s.xlsx' % wave_rate_date
             if os.path.exists(table_name):
-                logger.info('table %s already exist, try to delte and create a new one', table_name)
-                os.remove(table_name)
+                logger.info('table %s already exist, try to delete and create a new one', table_name)
+                try:
+                    os.remove(table_name)
+                except PermissionError as e:
+                    logger.error('export data error, file is open, please close file and retry export data')
+                    messagebox.showinfo('提示', '文件已打开, 无法导出文件, 请关闭文件后重新导出')
             with pd.ExcelWriter(table_name, engine='xlsxwriter') as writer:
                 def handle_output_format(sheet_name, df_inner):
                     worksheet = writer.sheets[sheet_name]
@@ -181,10 +210,10 @@ class DB:
                 workbook = writer.book
                 output_format = workbook.add_format({'bold': False, 'border': 0})
                 wave_rate_rank_col_names = ['timestamp', 'wave_rate_year', 'coin_type']
-                coin_rank_sheet = df[df['timestamp'] == max_date].copy()[wave_rate_rank_col_names]
+                coin_rank_sheet = df[df['timestamp'] == target_timestamp].copy()[wave_rate_rank_col_names]
                 coin_rank_sheet.sort_values(by='wave_rate_year', ascending=False, inplace=True)
                 coin_rank_sheet['rank'] = range(1, len(coin_rank_sheet) + 1)
-                coin_rank_sheet['timestamp'] = dt.datetime.fromtimestamp(max_date).strftime('%Y-%m-%d')
+                coin_rank_sheet['timestamp'] = wave_rate_date
                 coin_rank_sheet.rename(columns={**{k: v for k, v in self.wave_rate_col_name_map.items() if k in wave_rate_rank_col_names}, 'rank': '年华波动率排名'}, inplace=True)
                 coin_rank_sheet.to_excel(writer, sheet_name='年化波动率排名', index=False)
                 handle_output_format('年化波动率排名', coin_rank_sheet)
@@ -199,9 +228,57 @@ class DB:
             pass
         messagebox.showinfo('提示', '数据导出完成, 文件存放在%s目录下' % os.path.realpath('.'))
 
-    def update_wave_rate_date(self):
-        newest_timestamp = self.get_newest_date()
-        newest_data = data.get_kline_data(newest_timestamp)
+    def delete_dirty_wave_rate_data(self, timestamp):
+        logger.info('delete wave_rate data of date %s', dt.datetime.fromtimestamp(timestamp).isoformat())
+        self.execute('delete from wave_rate where timestamp>=%s' % timestamp)
+
+    def update_wave_rate_data(self):
+        logger.info('begin to periodically update wave rate data')
+        count, sleep_time = 1, 20
+        while count <= 5:
+            if self.state == 'initialing':
+                logger.info('storage is initializing, sleep %s and continue, retry time: %s', sleep_time, count)
+                time.sleep(sleep_time)
+                count += 1
+            elif self.state == 'ok':
+                break
+        start_timestamp = self.get_newest_date()
+        timestamp_delta = dt.datetime.now().timestamp() - start_timestamp
+        if timestamp_delta >= 24 * 3600 * self.data_clean_timespan:
+            self.clean_wave_rate_old_data()
+            self.init_wave_data()
+        else:
+            self.delete_dirty_wave_rate_data(start_timestamp)
+            newest_data = data.get_kline_data(before=start_timestamp * 1000 - 1)
+            old_data = self.execute_df('select * from wave_rate;')
+            old_coin_names = old_data['coin_type'].unique()
+            for coin_type in newest_data:
+                values = newest_data[coin_type]
+                if coin_type not in old_coin_names:
+                    self.handle_kline_data({coin_type: values})
+                else:
+                    values.sort(key=lambda x: int(x[0]))
+                    old_coin_data = old_data[old_data['coin_type'] == coin_type].sort_values(by='timestamp', ascending=True)
+                    wave_sum_window = deque([i for i in old_coin_data['daily_wave'].iloc[-14:]], maxlen=14)
+                    for index in range(len(values)):
+                        v = values[index]
+                        v[0] = int(v[0][:-3])
+                        v.append(coin_type)
+                        daily_wave = round(float(v[2]) - float(v[3]), ndigits=8)
+                        v.append(daily_wave)
+                        if len(wave_sum_window) < 14:
+                            wave_sum_window.append(daily_wave)
+                            continue
+                        else:
+                            wave_sum = sum(wave_sum_window)
+                            avg_wave = round(wave_sum / wave_coef, ndigits=8)
+                            daily_wave_rate = round(avg_wave / max(float(v[4]), 1e-8) * 100, ndigits=8)
+                            wave_rate_year = round(daily_wave_rate * wave_coef_year, ndigits=8)
+                            v.extend([avg_wave, wave_sum, daily_wave_rate, wave_rate_year])
+                            wave_sum_window.append(daily_wave)
+            self.insert_wave_rate_batch(newest_data)
+            self.clean_wave_rate_old_data()
+        logger.info('update start timestamp is %s', start_timestamp)
 
 
 db_inst = DB()
